pyproject.toml
[project]
name = "nexus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0,<5.1",
    "django-components>=0.28",
    "django-cors-headers>=4.3",
    "django-htmx>=1.18",
    "django-mptt>=0.15",
    "djangorestframework>=3.15",
    "faker>=40.1.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
